Fix readdir and implement seek for directory
If n <= 0, Readdir should return the remaining files and
never return io.EOF.
If n > 0, Readdir should only return io.EOF if it returns
an empty slice, per the documentation of os.Readdir.
It should also not reset the offset to 0.

Implement seek for directories when offset is 0 and
whence is io.SeekStart, because os.Seek supports this.